refactor: rename `getNewlinesErrors` -> `getNewlinesBetweenErrors`
